# frontend/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "llm-court API",
    "description": "LLM 法廷バトルゲームのバックエンド API(REST + SSE)。",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/sessions/{session_id}/events": {
      "get": {
        "tags": [
          "logs"
        ],
        "summary": "List Events",
        "description": "イベントログ(`after` より後の seq)。ディベート・裁判の両方で使う。",
        "operationId": "listEvents",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          },
          {
            "name": "after",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "After"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "oneOf": [
                      {
                        "$ref": "#/components/schemas/SessionStarted"
                      },
                      {
                        "$ref": "#/components/schemas/EvidenceCollected"
                      },
                      {
                        "$ref": "#/components/schemas/PhaseStarted"
                      },
                      {
                        "$ref": "#/components/schemas/StatementMade"
                      },
                      {
                        "$ref": "#/components/schemas/CitationIssuesDetected"
                      },
                      {
                        "$ref": "#/components/schemas/ClaimsExtracted"
                      },
                      {
                        "$ref": "#/components/schemas/JudgeScored"
                      },
                      {
                        "$ref": "#/components/schemas/VerdictDelivered"
                      },
                      {
                        "$ref": "#/components/schemas/SessionAborted"
                      },
                      {
                        "$ref": "#/components/schemas/ChoicesPrepared"
                      },
                      {
                        "$ref": "#/components/schemas/ChoiceMade"
                      },
                      {
                        "$ref": "#/components/schemas/PenaltyApplied"
                      },
                      {
                        "$ref": "#/components/schemas/TrialStarted"
                      },
                      {
                        "$ref": "#/components/schemas/TestimonyStarted"
                      },
                      {
                        "$ref": "#/components/schemas/TrialChoicesPrepared"
                      },
                      {
                        "$ref": "#/components/schemas/TrialChoiceMade"
                      },
                      {
                        "$ref": "#/components/schemas/WitnessResponded"
                      },
                      {
                        "$ref": "#/components/schemas/ContradictionSolved"
                      },
                      {
                        "$ref": "#/components/schemas/AnswerSubmitted"
                      },
                      {
                        "$ref": "#/components/schemas/TrialFinished"
                      },
                      {
                        "$ref": "#/components/schemas/ExplanationObjected"
                      },
                      {
                        "$ref": "#/components/schemas/LLMCallRecorded"
                      }
                    ],
                    "discriminator": {
                      "propertyName": "type",
                      "mapping": {
                        "session_started": "#/components/schemas/SessionStarted",
                        "evidence_collected": "#/components/schemas/EvidenceCollected",
                        "phase_started": "#/components/schemas/PhaseStarted",
                        "statement_made": "#/components/schemas/StatementMade",
                        "citation_issues_detected": "#/components/schemas/CitationIssuesDetected",
                        "claims_extracted": "#/components/schemas/ClaimsExtracted",
                        "judge_scored": "#/components/schemas/JudgeScored",
                        "verdict_delivered": "#/components/schemas/VerdictDelivered",
                        "session_aborted": "#/components/schemas/SessionAborted",
                        "choices_prepared": "#/components/schemas/ChoicesPrepared",
                        "choice_made": "#/components/schemas/ChoiceMade",
                        "penalty_applied": "#/components/schemas/PenaltyApplied",
                        "trial_started": "#/components/schemas/TrialStarted",
                        "testimony_started": "#/components/schemas/TestimonyStarted",
                        "trial_choices_prepared": "#/components/schemas/TrialChoicesPrepared",
                        "trial_choice_made": "#/components/schemas/TrialChoiceMade",
                        "witness_responded": "#/components/schemas/WitnessResponded",
                        "contradiction_solved": "#/components/schemas/ContradictionSolved",
                        "answer_submitted": "#/components/schemas/AnswerSubmitted",
                        "trial_finished": "#/components/schemas/TrialFinished",
                        "explanation_objected": "#/components/schemas/ExplanationObjected",
                        "llm_call_recorded": "#/components/schemas/LLMCallRecorded"
                      }
                    }
                  },
                  "title": "Response Listevents"
                }
              }
            }
          },
          "404": {
            "description": "セッションがない",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "現在の状態ではその操作をできない",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/sessions/{session_id}/stream": {
      "get": {
        "tags": [
          "logs"
        ],
        "summary": "Stream Session",
        "operationId": "streamSession",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          },
          {
            "name": "after",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "After"
            }
          },
          {
            "name": "last-event-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Last-Event-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "SSE。event は debate(永続イベント、id=seq)/ turn(発言開始)/ token(発言のチャンク)/ progress / task。裁判では turn の代わりに witness(証人の応答開始)を送る。判決・閉廷・中断で終わる",
            "content": {
              "text/event-stream": {}
            }
          },
          "404": {
            "description": "セッションがない",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "現在の状態ではその操作をできない",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/cases": {
      "get": {
        "tags": [
          "trials"
        ],
        "summary": "List Cases",
        "description": "遊べる事件の一覧(公開の情報だけ)。既定では solver の検証に合格した事件だけ。",
        "operationId": "listCases",
        "parameters": [
          {
            "name": "all",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "解けると判定されていない事件も含める",
              "default": false,
              "title": "All"
            },
            "description": "解けると判定されていない事件も含める"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CaseSummary"
                  },
                  "title": "Response Listcases"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/trials": {
      "post": {
        "tags": [
          "trials"
        ],
        "summary": "Create Trial",
        "description": "事件を選んで開廷する。最初の証言の選択肢まで用意して返す。",
        "operationId": "createTrial",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTrialRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrialView"
                }
              }
            }
          },
          "404": {
            "description": "裁判・事件がない",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "現在の状態ではその操作をできない",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/trials/{session_id}": {
      "get": {
        "tags": [
          "trials"
        ],
        "summary": "Get Trial",
        "operationId": "getTrial",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrialView"
                }
              }
            }
          },
          "404": {
            "description": "裁判・事件がない",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "現在の状態ではその操作をできない",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/trials/{session_id}/advance": {
      "post": {
        "tags": [
          "trials"
        ],
        "summary": "Advance Trial",
        "description": "選択肢の用意待ちなら用意する(通常は選択のたびに自動で用意される)。",
        "operationId": "advanceTrial",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskAccepted"
                }
              }
            }
          },
          "404": {
            "description": "裁判・事件がない",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "現在の状態ではその操作をできない",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/trials/{session_id}/choices": {
      "post": {
        "tags": [
          "trials"
        ],
        "summary": "Choose Trial Option",
        "description": "選択肢を選ぶ。証人の応答を SSE でストリーミングし、次の選択肢(または問い)まで進める。",
        "operationId": "chooseTrialOption",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChooseRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskAccepted"
                }
              }
            }
          },
          "404": {
            "description": "裁判・事件がない",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "現在の状態ではその操作をできない",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "選択肢・項目がない",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/trials/{session_id}/answer": {
      "post": {
        "tags": [
          "trials"
        ],
        "summary": "Answer Trial",
        "description": "最後の問いに答えて閉廷する。閉廷後は解説を返す。",
        "operationId": "answerTrial",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnswerRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrialView"
                }
              }
            }
          },
          "404": {
            "description": "裁判・事件がない",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "現在の状態ではその操作をできない",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "選択肢・項目がない",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/trials/{session_id}/objections": {
      "post": {
        "tags": [
          "trials"
        ],
        "summary": "Object To Explanation",
        "description": "「解説に異議あり」。解説の項目の誤りを報告する(閉廷後のみ)。",
        "operationId": "objectToExplanation",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ObjectionRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrialView"
                }
              }
            }
          },
          "404": {
            "description": "裁判・事件がない",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "現在の状態ではその操作をできない",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "選択肢・項目がない",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "EvidenceCollected": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id",
            "default": ""
          },
          "seq": {
            "type": "integer",
            "title": "Seq",
            "default": 0
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          },
          "role": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Role"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "prompt_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt Version"
          },
          "type": {
            "type": "string",
            "const": "evidence_collected",
            "title": "Type",
            "default": "evidence_collected"
          },
          "report": {
            "$ref": "#/components/schemas/ResearchReport"
          }
        },
        "type": "object",
        "required": [
          "report"
        ],
        "title": "EvidenceCollected"
      },
...
      "JudgeScored": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id",
            "default": ""
          },
          "seq": {
            "type": "integer",
            "title": "Seq",
            "default": 0
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          },
          "role": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Role"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "prompt_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt Version"
          },
          "type": {
            "type": "string",
            "const": "judge_scored",
            "title": "Type",
            "default": "judge_scored"
          },
          "score": {
            "$ref": "#/components/schemas/JudgeScore"
          }
        },
        "type": "object",
        "required": [
          "score"
        ],
        "title": "JudgeScored"
      },
...
      "PhaseStarted": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id",
            "default": ""
          },
          "seq": {
            "type": "integer",
            "title": "Seq",
            "default": 0
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          },
          "role": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Role"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "prompt_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt Version"
          },
          "type": {
            "type": "string",
            "const": "phase_started",
            "title": "Type",
            "default": "phase_started"
          },
          "phase": {
            "$ref": "#/components/schemas/DebatePhase"
          },
          "round": {
            "type": "integer",
            "title": "Round",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "phase"
        ],
        "title": "PhaseStarted"
      },
...
      "SessionStarted": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id",
            "default": ""
          },
          "seq": {
            "type": "integer",
            "title": "Seq",
            "default": 0
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          },
          "role": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Role"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "prompt_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt Version"
          },
          "type": {
            "type": "string",
            "const": "session_started",
            "title": "Type",
            "default": "session_started"
          },
          "topic": {
            "type": "string",
            "title": "Topic"
          },
          "mode": {
            "type": "string",
            "title": "Mode"
          },
          "rounds": {
            "type": "integer",
            "title": "Rounds"
          },
          "models": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Models"
          },
          "human_side": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Side"
              },
              {
                "type": "null"
              }
            ]
          },
          "penalty_gauge": {
            "type": "integer",
            "title": "Penalty Gauge",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "topic",
          "mode",
          "rounds",
          "models"
        ],
        "title": "SessionStarted"
      },
...
      "StatementMade": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id",
            "default": ""
          },
          "seq": {
            "type": "integer",
            "title": "Seq",
            "default": 0
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          },
          "role": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Role"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "prompt_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt Version"
          },
          "type": {
            "type": "string",
            "const": "statement_made",
            "title": "Type",
            "default": "statement_made"
          },
          "statement": {
            "$ref": "#/components/schemas/Statement"
          },
          "call_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Call Id"
          }
        },
        "type": "object",
        "required": [
          "statement"
        ],
        "title": "StatementMade"
      },
...
    }
  }
}